修复有些版本的编译器会认为枚举类型是signed int而导致的warning

+++ src/log/log_sink_file_backend.cpp
@@ -86,7 +86,7 @@ namespace util {
 
             f->write(content, content_size);
             f->put('\n');
-            if (caller.level_id <= log_file_.auto_flush) {
+            if (static_cast<uint32_t>(caller.level_id) <= log_file_.auto_flush) {
                 f->flush();
             }
